Add plastic emission calculation and update related functions for emission factors

+++ my_microplastic_sim/exporter.py
@@ -1,3 +1,4 @@
+import os
 import pandas as pd
 #from openpyxl import load_workbook
 from PA.pa import pa
@@ -5,15 +6,15 @@ from PET.pet import pet
 from PS.ps import ps
 from processing.em_factor import calculate_emission_factors
 
-def save_results_to_excel(input_file, output_file, all_results):
+def save_results_to_excel(output_file, all_results):
     data = []
     
-    pa_df = pa()
-    pet_df = pet()
-    ps_df = ps()
+    pa_df,_,_= pa()
+    pet_df,_,_ = pet()
+    ps_df,_,_ = ps()
 
     # Calculate emission factors
-    emission_factors = calculate_emission_factors()
+    emission_factors,_,_ = calculate_emission_factors()
     emission = []
 
     for water_type, polymers_data in all_results.items():
@@ -41,28 +42,22 @@ def save_results_to_excel(input_file, output_file, all_results):
     try:
         df = pd.DataFrame(data)
         emission_df = pd.DataFrame(emission)
-        with pd.ExcelWriter(output_file, engine='openpyxl', mode='a', if_sheet_exists='replace') as writer:
+        if os.path.exists(output_file):
+            # File exists: append/replace sheets
+            with pd.ExcelWriter(output_file, engine='openpyxl', mode='a', if_sheet_exists='replace') as writer:
+                    df.to_excel(writer, sheet_name='Polymer_Results', index=False)
+                    pa_df.to_excel(writer, sheet_name="PA", index=False)
+                    pet_df.to_excel(writer, sheet_name="PET", index=False)
+                    ps_df.to_excel(writer, sheet_name="PS", index=False)
+                    emission_df.to_excel(writer, sheet_name="Emission_Factors", index=False)
+        else:
+            # File doesn't exist: create and write all sheets
+            with pd.ExcelWriter(output_file, engine='openpyxl') as writer:
                 df.to_excel(writer, sheet_name='Polymer_Results', index=False)
                 pa_df.to_excel(writer, sheet_name="PA", index=False)
                 pet_df.to_excel(writer, sheet_name="PET", index=False)
                 ps_df.to_excel(writer, sheet_name="PS", index=False)
                 emission_df.to_excel(writer, sheet_name="Emission_Factors", index=False)
-        '''
-        # if input_file == output_file:
-        #     with pd.ExcelWriter(output_file, engine='openpyxl', mode='a', if_sheet_exists='replace') as writer:
-        #         df.to_excel(writer, sheet_name='Polymer_Results', index=False)
-        #         pa_df.to_excel(writer, sheet_name="PA", index=False)
-        #         pet_df.to_excel(writer, sheet_name="PET", index=False)
-        #         ps_df.to_excel(writer, sheet_name="PS", index=False)
-        #         emission_df.to_excel(writer, sheet_name="Emission_Factors", index=False)
-        # else:
-        #     with pd.ExcelWriter(output_file, engine='openpyxl', mode='a', if_sheet_exists='replace') as writer:
-        #         df.to_excel(writer, sheet_name='Polymer_Results', index=False)
-        #         pa_df.to_excel(writer, sheet_name="PA", index=False)
-        #         pet_df.to_excel(writer, sheet_name="PET", index=False)
-        #         ps_df.to_excel(writer, sheet_name="PS", index=False)
-        #         emission_df.to_excel(writer, sheet_name="Emission_Factors", index=False)
-        '''
 
     except Exception as e:
         print(f"Error saving results to Excel: {e}")
